+++ OneBlink.SDK/models/FormsAppBase.cs
@@ -72,5 +72,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public bool enableAppUserSignup
+        {
+            get; set;
+        }
     }
 }