+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 -   `searchQuerystringParameter` to `FormElement` class
+-   `personalisation` and `externalIdGenerationOnSubmit` to `Form` class
+-   `FormPersonalisation` model class for `personalisation`
 
 ## 4.1.8 (2023-05-12)