+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   `searchQuerystringParameter` to `FormElement` class
 -   `personalisation` and `externalIdGenerationOnSubmit` to `Form` class
 -   `FormPersonalisation` model class for `personalisation`
+-   `enableAppUserSignup` to `FormsAppBase` class
 
 ## 4.1.8 (2023-05-12)