+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 -   `personalisation` and `externalIdGenerationOnSubmit` to `Form` class
 -   `FormPersonalisation` model class for `personalisation`
 -   `enableAppUserSignup` to `FormsAppBase` class
+-   `firstName` and `lastName` to `FormsAppUser`
 
 ## 4.1.8 (2023-05-12)